moved tournament components
