+++ src/frontend/app/pages/Tournament/components/TournamentCard.tsx
