+++ src/frontend/app/pages/Tournament/components/TournamentSection.tsx
